requirements.txt
Django>=4.0.4,<4.1
djangorestframework>=3.13.1,<3.14
psycopg2>=2.9.3,<2.10
drf-spectacular>=0.22.1,<0.23
Pillow>=9.1.0,<9.2
django-countries>=7.3.2,<7.4
django-ckeditor>=6.5.1,<6.6
requests==2.28.1